First steps firebase

+++ src/lib/functions/postsData.js
@@ -4,7 +4,7 @@ import { editPost }  from '../components/modalPostEdit.js'
 export const savePost = () => {
   const createdPost = document.getElementById('writtePost').value;
   if (createdPost !== '') {
-    return firebase.firestore()
+      return firebase.firestore()
       .collection('posts')
       .add({
         userId: firebase.auth().currentUser.uid,
@@ -15,6 +15,9 @@ export const savePost = () => {
         postDislikes :[],
         // Agregar imagenes, etiquetas
         timestamp: firebase.firestore.FieldValue.serverTimestamp()
+      })
+      .then(()=>{
+        window.location.reload();
       })
       .catch(function (error) {
         console.error('Error writing new message to database', error);
@@ -45,12 +48,12 @@ export const showUserPosts = () => {
       let arrayUserPosts = doc.data();
       const userPostsTemplate = `<div class='postContainer' value="${doc.id}">
                                 <div class='post'>
-                                <div class='postUserphoto'><img class='postUserphoto' src="${firebase.auth().photoUrl}"></div>
+                                <div class='postUserphoto'></div>
                                 <div class='postInfo'>
                                 <h2 class='postedUsername'>${arrayUserPosts.userName}
                                 <i id ='editPost' class="fas fa-ellipsis-h">
                                 <div id='menuEdit' class="menuEdit">
-                                    <button id = 'editpost' value='${doc.id}'>Editar</button></a>
+                                    <button id = 'editPostButton' value='${doc.id}'>Editar</button></a>
                                     <button id ='deletepost' value='${doc.id}'>Borrar</button>
                                  </div>
                                  </i>
@@ -69,7 +72,7 @@ export const showUserPosts = () => {
                                 </div>`;
     userPosts.innerHTML += userPostsTemplate;
   });  
-    const clickEdit = document.querySelectorAll('#editpost');
+    const clickEdit = document.querySelectorAll('#editPostButton');
     clickEdit.forEach(item => {item.addEventListener('click', () => editPost (item.value))});
     const clickDelete = document.querySelectorAll('#deletepost');
     clickDelete.forEach(item => {item.addEventListener('click', () => deletePost (item.value))});
@@ -89,10 +92,14 @@ export const showUserPosts = () => {
 };
 let showMenuEditcontrol = true;
 
-const deletePost= (id) => {
-    firebase.firestore().collection("posts").doc(id).delete().then(() => {
-    console.log("Document successfully deleted! " + id);
-}).catch((error) => {
-  console.error("Error removing document: ", error);
+const deletePost = (id) => {
+  console.log('entra a delete' + id);
+    if (confirm("Segura que quieres borrar este post!")) {
+      firebase.firestore().collection("posts").doc(id).delete().then(() => {
+        console.log("Document successfully deleted! " + id);
+        window.location.reload()
+    }).catch((error) => {
+      console.error("Error removing document: ", error);
+    })
 }
-)};
+};
